feat(icon): merge custom configuration with the default one instead of overriding it
Previously, a custom configuration would completely override the default one. Now only the settings
specified in the custom configuration will be overridden.

OUI-2266
BREAKING CHANGE:
When passing a custom configuration, settings that are not specified within it will not be
overridden anymore. Projects that rely on this behavior to override default settings will need to
specify all settings in the custom configuration.